+++ replicaset.go
@@ -83,7 +83,7 @@ func (rs *Replicaset) BucketStat(ctx context.Context, bucketID uint64) (BucketSt
 func (rs *Replicaset) bucketStatAsync(ctx context.Context, bucketID uint64) *tarantool.Future {
 	const bucketStatFnc = "vshard.storage.bucket_stat"
 
-	return rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.RO}, bucketStatFnc, []interface{}{bucketID})
+	return rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.RO}, bucketStatFnc, []any{bucketID})
 }
 
 type vshardStorageBucketStatResponseProto struct {
@@ -154,13 +154,13 @@ func bucketStatWait(future *tarantool.Future) (BucketStatInfo, error) {
 }
 
 // CallAsync sends async request to remote storage
-func (rs *Replicaset) CallAsync(ctx context.Context, opts ReplicasetCallOpts, fnc string, args interface{}) *tarantool.Future {
+func (rs *Replicaset) CallAsync(ctx context.Context, opts ReplicasetCallOpts, fnc string, args any) *tarantool.Future {
 	if opts.Timeout > 0 {
 		// Don't set any timeout by default, parent context timeout would be inherited in this case.
 		// Don't call cancel in defer, because this we send request asynchronously,
 		// and wait for result outside from this function.
-		// suppress linter warning: lostcancel: the cancel function returned by context.WithTimeout should be called, not discarded, to avoid a context leak (govet)
-		//nolint:govet
+		// suppress linter warning: lostcancel: the cancel function returned by context.WithTimeout should be called, not discarded, to avoid a context leak (govet, gosec G118)
+		//nolint:govet,gosec
 		ctx, _ = context.WithTimeout(ctx, opts.Timeout)
 	}
 
@@ -179,7 +179,7 @@ func (rs *Replicaset) bucketsDiscoveryAsync(ctx context.Context, from uint64) *t
 	}{From: from}
 
 	return rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.PreferRO}, bucketsDiscoveryFnc,
-		[]interface{}{bucketsDiscoveryPaginationRequest})
+		[]any{bucketsDiscoveryPaginationRequest})
 }
 
 type bucketsDiscoveryResp struct {
@@ -192,7 +192,7 @@ func bucketsDiscoveryWait(future *tarantool.Future) (bucketsDiscoveryResp, error
 	// https://github.com/tarantool/vshard/blob/8d299bfecff8bc656056658350ad48c829f9ad3f/vshard/router/init.lua#L343
 	var resp bucketsDiscoveryResp
 
-	err := future.GetTyped(&[]interface{}{&resp})
+	err := future.GetTyped(&[]any{&resp})
 	if err != nil {
 		return resp, fmt.Errorf("future.GetTyped() failed: %v", err)
 	}
@@ -297,7 +297,7 @@ func (rs *Replicaset) BucketsCount(ctx context.Context) (uint64, error) {
 	var bucketCount uint64
 
 	fut := rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.ANY}, bucketCountFnc, nil)
-	err := fut.GetTyped(&[]interface{}{&bucketCount})
+	err := fut.GetTyped(&[]any{&bucketCount})
 
 	return bucketCount, err
 }
@@ -305,7 +305,7 @@ func (rs *Replicaset) BucketsCount(ctx context.Context) (uint64, error) {
 func (rs *Replicaset) BucketForceCreate(ctx context.Context, firstBucketID, count uint64) error {
 	const bucketForceCreateFnc = "vshard.storage.bucket_force_create"
 
-	fut := rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.RW}, bucketForceCreateFnc, []interface{}{firstBucketID, count})
+	fut := rs.CallAsync(ctx, ReplicasetCallOpts{PoolMode: pool.RW}, bucketForceCreateFnc, []any{firstBucketID, count})
 	_, err := fut.GetResponse()
 
 	return err